feat: add read status property to book object
Extended Book constructor to include a read status property. The value
is fetched from the checkbox input in the dialog form and stored as a
boolean in the object. Prepares for upcoming feature to toggle read
status. The read status value of the sample books are pre set in their
object itself.